documented environment
